🚚 Restructure trait architecture
