+++ src/ecs/traits/index.ts
@@ -1,0 +1,6 @@
+export * from "./tags";
+export * from "./movement";
+export * from "./pheromones";
+export * from "./refs";
+export * from "./relations";
+export * from "./tags";
